refactor: rename ingredientsDataAPI to stellarBurgersAPI
